% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn.nsi-package.R
\docType{package}
\name{bn.nsi}
\alias{bn.nsi}
\title{bn.nsi: A package for loading, saving and passing to function project settings}
\description{
Interface for loading, saving and passing to function project settings
}
\section{Main functions for operate NSI}{

\link{is.nsi}, \link{nsi.params}, \link{nsi.global}, \link{is.nsi_loaded}, \link{nsi.load}
\link{nsi.read}, \link{nsi.save}, \link{nsi.init}
}

\section{Utility functions for operate files}{

\link{fdata}, \link{frdata}, \link{fgraph}, \link{freport}, \link{fdict}.
}

\section{Misc functions}{

\link{rflat}
}

\author{
Rinat Levchuk (bane.tsk@gmail.com)
}
