% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn.nsi.R
\name{fpath.funcs}
\alias{fpath.funcs}
\alias{fpath}
\alias{fdata}
\alias{frdata}
\alias{fgraph}
\alias{freport}
\alias{fdict}
\title{Utility functions for operate files}
\usage{
fpath(nsi.dir, fname)

fdata(fname)

frdata(fname)

fgraph(fname)

freport(fname)

fdict(fname)
}
\arguments{
\item{nsi.dir}{A desired directory from NSI}

\item{fname}{A filename to glue with}
}
\description{
Utility functions for operate files
}
\examples{
## generic
fpath("data", "dataset.xlsx")
## rendered graphics
fgraph("plot.png")
}
\concept{functions for operate files}
