% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn.nsi.R
\name{is.nsi}
\alias{is.nsi}
\title{Check what object is a NSI structure}
\usage{
is.nsi(nsi)
}
\arguments{
\item{nsi}{an NSI structure}
}
\value{
TRUE if \code{nsi} is an NSI structure or FALSE if not
}
\description{
\code{is.nsi} returns TRUE if object is a NSI structure
}
\details{
\code{is.nsi(nsi)} ensures that \code{nsi} is a list and then check for this fields in \code{nsi}:
\describe{
 \item{tag}{a unique tag needed to distinct one nsi from another}
 \item{dir}{a list with data, results, graphs, reports, dict fields}
}
}
\seealso{
nsi-template
}
