% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn.nsi.R
\name{nsi-template}
\alias{nsi-template}
\title{NSI struct template}
\format{
NSI is an list() and may contains at least this elements:
\describe{
  \item{tag}{Unique, amoung yours projects, tag with version}
  \item{data}{A path to input datasets}
  \item{dirs}{A list() with subpaths}
  \item{...}{Any records of any types you want}
}
\code{dirs} fields are follow:
\describe{
  \item{data}{A path to input datasets, mandatory}
  \item{results}{A path for all outputs produced, mandatory}
  \item{rdata}{A path for resulting datasets produced, optional}
  \item{graphs}{A path for graphics rendered, mandatory}
  \item{reports}{A path for store generated reports, , optional. Usual is same as \code{results}}
  \item{dicts}{A path for loading dictionaries, optional (if any)}
}
}
\description{
NSI struct template
}
