% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn.nsi.R
\name{nsi.init}
\alias{nsi.init}
\title{Create NSI structure}
\usage{
nsi.init(tag, file, optional = FALSE)
}
\arguments{
\item{tag}{A tag for project}

\item{file}{An name to JSON file}

\item{optional}{if TRUE - use NSI with optional fields}
}
\value{
TRUE if success
}
\description{
Create NSI structure from inner template
Template can be with optional fields
}
