% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn.nsi.R
\name{nsi.load}
\alias{nsi.load}
\title{Load NSI from file}
\usage{
nsi.load(file, reload = FALSE)
}
\arguments{
\item{file}{An name to JSON file, in case success load - will be saved to .nsi.fn}

\item{reload}{if FALSE (default) - do not overwrite nsi.globals}
}
\value{
TRUE if success
}
\description{
Load NSI from \code{file} (default) to nsi.global or reload
}
