% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn.nsi.R
\name{nsi.params}
\alias{nsi.params}
\title{Wrapper function to check & load nsi from param or from nsi.global}
\usage{
nsi.params(nsi = NULL, global = TRUE, check = FALSE)
}
\arguments{
\item{nsi}{an NSI structure or NULL to load nsi.global}

\item{global}{If TRUE - try return nsi.global}

\item{check}{if FALSE - suppress checking (do this in stable code)}
}
\value{
NSI structure
}
\description{
\code{nsi.params} intended for use inside functions
if \code{nsi} is correct NSI - check (if needed) and simple returns it
if \code{nsi == NULL} - return nsi.global or fail if not exists
checks may be suppressed
}
\seealso{
nsi-template
}
