% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn.nsi.R
\name{nsi.read}
\alias{nsi.read}
\title{Read NSI from file and return}
\usage{
nsi.read(file)
}
\arguments{
\item{file}{An name to JSON file}
}
\value{
NSI if success
}
\description{
Used in case modification of NSI
Just read NSI from JSON and returns it
}
