% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn.nsi.R
\name{nsi.save}
\alias{nsi.save}
\title{Save NSI to JSON file}
\usage{
nsi.save(file = NULL, nsi = NULL, force = FALSE)
}
\arguments{
\item{file}{An name to JSON file}

\item{nsi}{if NULL - saving .nsi.global to file, else - save given nsi}

\item{force}{Overwrite JSON file}
}
\value{
TRUE if seccess
}
\description{
Used in case modification of NSI
Just read NSI from JSON and returns it
}
