% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn.nsi.R
\name{rflat}
\alias{rflat}
\title{Make vector from data.frame row}
\usage{
rflat(row)
}
\arguments{
\item{row}{A row from data.frame / tibble}
}
\value{
raw vector
}
\description{
Make vector from data.frame row
}
